Rearrange &run so it can capture output for win32
Why? Because win32 can only select on sockets. So we have to do some
convincing to make it happen. First, instead of using the useful geniosym,
we must use socketpair. Second, open3 will by default use pipe given a
handle, so we have convince open3 to dup the file handle instead. We do
that by using the string form of open3 and the fileno of the handle.
After that, win32 will do the right thing, but don't deviate, or none
of it will work.