Make sure that the inflation produced something
So it appears that if gunzip errors out, producing 0 byte output, the tar
from libarchive will not error out like gnu tar. So, from a pipe
perspective, there was no error, but there was also no output. This is in
place to prevent that, and other like errors, from occuring.